refactor: merge `EmptyState`s into one component
