--- src/modules/Dashboard/features/Participant/Assignments/AboutParticipant/EmptyState/EmptyState.types.ts
@@ -1,4 +1,0 @@
-export type EmptyStateProps = {
-  onClickAssign: () => void;
-  isTeamMember: boolean;
-};
