--- src/modules/Dashboard/features/Participant/Assignments/ByParticipant/EmptyState/EmptyState.types.ts
@@ -1,4 +1,0 @@
-export type EmptyStateProps = {
-  onClickAssign: () => void;
-  isLimitedAccount: boolean;
-};
