--- src/modules/Dashboard/features/Participant/Assignments/ByParticipant/EmptyState/index.ts
@@ -1,1 +1,0 @@
-export * from './EmptyState';
