+++ src/modules/Dashboard/features/Participant/Assignments/EmptyState/EmptyState.tsx
@@ -6,18 +6,18 @@ import { StyledFlexAllCenter, StyledFlexColumn, StyledHeadline, variables } from
 
 import { EmptyStateProps } from './EmptyState.types';
 
-export const EmptyState = ({ onClickAssign, isTeamMember }: EmptyStateProps) => {
+export const EmptyState = ({ icon, title, onClickAssign }: EmptyStateProps) => {
   const { t } = useTranslation('app', { keyPrefix: 'participantDetails' });
 
   return (
     <StyledFlexAllCenter sx={{ flexDirection: 'column', flex: 1, m: 'auto', textAlign: 'center' }}>
       <StyledFlexColumn sx={{ alignItems: 'center', gap: 1.6, maxWidth: '50.7rem' }}>
-        <Svg id="about-participant" width="80" height="80" fill={variables.palette.outline} />
+        <Svg id={icon} width="80" height="80" fill={variables.palette.outline} />
         <StyledHeadline as="h2" sx={{ color: variables.palette.outline, m: 0 }}>
-          {isTeamMember ? t('aboutParticipantEmptyTeamMember') : t('aboutParticipantEmpty')}
+          {title}
         </StyledHeadline>
       </StyledFlexColumn>
-      {!isTeamMember && (
+      {!!onClickAssign && (
         <Button variant="contained" color="primary" onClick={onClickAssign} sx={{ mt: 2.4 }}>
           {t('assignActivityButton')}
         </Button>
