+++ src/modules/Dashboard/features/Participant/Assignments/EmptyState/EmptyState.types.ts
@@ -1,0 +1,7 @@
+import { Icons } from 'svgSprite';
+
+export type EmptyStateProps = {
+  icon: Icons;
+  title: string;
+  onClickAssign?: () => void;
+};
